correccion final, se corrige la carga de las propiedasde de monto de tarifa y se corrige el calculo de duracio de visita.

+++ DSI2021/Frontend/PantallaRegistroVenta.cs
@@ -26,8 +26,8 @@ namespace DSI2021.Frontend {
         }
 
         private void btnGenerar(object sender, EventArgs e) {
-            int montoPorEntrada;
-            int montoTotal;
+            decimal montoPorEntrada;
+            decimal montoTotal;
 
             cantidadEntradas = GetCantidadEntradas();
